+++ packages/developer_tools_network/lib/network_inspector/model/network_translation.dart
@@ -74,6 +74,8 @@ enum NetworkTranslationKey {
   callsListDelete,
   callsListStats,
   callsListSave,
+  callsListExportPostman,
+  callsListCopyPostman,
   logsEmpty,
   logsError,
   logsItemError,